Fix a subtle bug with flatMap and related streams
If a new parent was added to the flatMapped stream, and the new parent
was updated on the same tick, and some other parent was also updated on
the same tick, then the other parent took precedence. Now the new one
does. In practice, this would have never happened. But.